requirements.txt
tabulate
PyMuPDF
pandas